Move setupIPv6Forwarding() to setup_ip_forwarding.go
